backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
emergentintegrations --extra-index-url https://d33sy5i8bnduwe.cloudfront.net/simple/